+++ src/core/components/amdutils.cpp
@@ -4,11 +4,11 @@
 #include "amdutils.h"
 
 #include "common/stringutils.h"
-#include "units/units.h"
 #include <algorithm>
 #include <cstddef>
 #include <iterator>
 #include <regex>
+#include <units.h>
 #include <unordered_map>
 #include <utility>