+++ src/core/components/controls/amd/pm/advanced/overdrive/freqvolt/pmfreqvolt.h
@@ -4,12 +4,12 @@
 #pragma once
 
 #include "core/components/controls/control.h"
-#include "units/units.h"
 #include <map>
 #include <memory>
 #include <string>
 #include <string_view>
 #include <tuple>
+#include <units.h>
 #include <unordered_map>
 #include <utility>
 #include <vector>