+++ src/core/components/controls/amd/pm/handlers/ppdpmhandler.h
@@ -4,9 +4,9 @@
 #pragma once
 
 #include "ippdpmhandler.h"
-#include "units/units.h"
 #include <memory>
 #include <string>
+#include <units.h>
 #include <utility>
 #include <vector>