+++ src/core/components/sensors/amd/fanspeedrpm.cpp
@@ -17,12 +17,12 @@
 #include "core/profilepartprovider.h"
 #include "core/profilepartxmlparserprovider.h"
 #include "core/sysfsdatasource.h"
-#include "units/units.h"
 #include <filesystem>
 #include <memory>
 #include <optional>
 #include <string>
 #include <tuple>
+#include <units.h>
 #include <utility>
 #include <vector>