+++ src/core/components/sensors/amd/gputemp.cpp
@@ -18,13 +18,13 @@
 #include "core/profilepartxmlparserprovider.h"
 #include "core/sysfsdatasource.h"
 #include "easyloggingpp/easylogging++.h"
-#include "units/units.h"
 #include <filesystem>
 #include <fmt/format.h>
 #include <memory>
 #include <optional>
 #include <string>
 #include <tuple>
+#include <units.h>
 #include <utility>
 #include <vector>