+++ src/core/components/sensors/amd/memusage.cpp
@@ -18,13 +18,13 @@
 #include "core/iprofilepartxmlparser.h"
 #include "core/profilepartprovider.h"
 #include "core/profilepartxmlparserprovider.h"
-#include "units/units.h"
 #include <cstdint>
 #include <filesystem>
 #include <memory>
 #include <optional>
 #include <string>
 #include <tuple>
+#include <units.h>
 #include <utility>
 #include <vector>