+++ src/core/components/sensors/cpu/cpufreqpack.cpp
@@ -17,13 +17,13 @@
 #include "core/profilepartxmlparserprovider.h"
 #include "core/sysfsdatasource.h"
 #include "easyloggingpp/easylogging++.h"
-#include "units/units.h"
 #include <algorithm>
 #include <filesystem>
 #include <fmt/format.h>
 #include <memory>
 #include <optional>
 #include <string>
+#include <units.h>
 #include <utility>
 #include <vector>