+++ src/core/components/sensors/sensor.h
@@ -5,13 +5,13 @@
 
 #include "core/components/sensors/isensor.h"
 #include "core/idatasource.h"
-#include "units/units.h"
 #include <atomic>
 #include <functional>
 #include <memory>
 #include <optional>
 #include <string>
 #include <string_view>
+#include <units.h>
 #include <utility>
 #include <vector>