+++ tests/src/test_amdpmvoltoffset.cpp
@@ -8,7 +8,7 @@
 #include "common/stringdatasourcestub.h"
 #include "common/vectorstringdatasourcestub.h"
 #include "core/components/controls/amd/pm/advanced/overdrive/voltoffset/pmvoltoffset.h"
-#include "units/units.h"
+#include <units.h>
 
 extern template struct trompeloeil::reporter<trompeloeil::specialized>;