+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Prefer external third-party libraries over the bundled ones. This should help package maintainers to follow their distribution policies (#13, #346). Version requirements for the external libraries:
   - fmt 5.0 or later.
   - pugixml 1.11 or later.
+  - units 2.3.1 or later.
 - Replace file headers license section with [SPDX unique license identifiers](https://spdx.dev/ids/).
 
 ### Added